Feature tests added

+++ app/Services/TaskService.php
@@ -3,6 +3,7 @@
 namespace App\Services;
 
 use App\Dto\TaskDto;
+use App\Exceptions\TaskException;
 use App\Models\Task;
 use App\Repositories\Interfaces\TaskRepositoryInterface;
 use App\Services\Interfaces\TaskServiceInterface;
@@ -35,6 +36,8 @@ class TaskService implements TaskServiceInterface
 
     public function save(TaskDto $dto): Task
     {
+        // automatically set status = todo
+
         return $this->taskRepository->save(
             $dto->getName(),
             $dto->getDescription(),
@@ -42,17 +45,42 @@ class TaskService implements TaskServiceInterface
         );
     }
 
+    /**
+     * @throws TaskException
+     */
     public function update(TaskDto $dto, $id): Task
     {
+        /**
+         * task statuses: done, in_progress, todo
+         *
+         *  if task status === done - task can not be updated. Throw exception
+         */
+
+        // find task by id
+        // check task status
+        // if task status === done - task can not be updated -> Throw exception
+        // update task
+
+        $checkTaskStatus = $this->taskRepository->checkTaskStatus($id);
+        if ($checkTaskStatus->status === 'done'){
+            throw new TaskException();
+        }
         return $this->taskRepository->update(
             $dto->getName(),
             $dto->getDescription(),
             $dto->getStatus(),
-            $id);
+            $id
+        );
     }
 
     public function delete($id): void
     {
-        $this->taskRepository->delete($id);
+        // only task with status "done" can be deleted
+        $task = $this->taskRepository->checkTaskStatus($id);
+        if ($task->status == 'done'){
+            return;
+        } else {
+            $this->taskRepository->delete($id);
+        }
     }
 }
